requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anndata==0.12.10
annotated-types==0.7.0
appnope==0.1.4
array-api-compat==1.14.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
cytetype==0.19.2
debugpy==1.8.20
decorator==5.2.1
docrep==0.3.2
donfig==0.8.1.post1
duckdb==1.4.4
executing==2.2.1
fast-array-utils==1.3.1
filelock==3.25.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-crc32c==1.8.0
grpcio==1.78.0
h5py==3.15.1
hdf5plugin==6.0.0
idna==3.11
igraph==1.0.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
legacy-api-wrap==1.5
leidenalg==0.11.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_collections==1.1.0
mpmath==1.3.0
mudata==0.3.3
multidict==6.7.1
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numcodecs==0.16.5
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opt_einsum==3.4.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
scanpy==1.12
scikit-learn==1.8.0
scikit-misc==0.5.2
scipy==1.17.1
scvi-tools==1.4.2
seaborn==0.13.2
session-info2==0.4
session_info==1.0.1
setuptools==82.0.0
six==1.17.0
sparse==0.18.0
stack-data==0.6.3
statsmodels==0.14.6
stdlib-list==0.12.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.6.0
torch==2.10.0
torchmetrics==1.8.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.6
xarray==2026.2.0
yarl==1.23.0
zarr==3.1.5
